steam部分游戏(pubg)战绩采集程序

+++ src/record/pubg/match.go
@@ -21,35 +21,35 @@ type PlayerData struct {
 	} `json:"LiveTracking"bson:"live_tracking"`
 	MatchHistory []struct {
 		Assists              int     `json:"Assists"bson:"assists"`
-		Damage               int     `json:"Damage"bson:"damage"`
-		Headshots            int     `json:"Headshots"bson:"headshots"`
+		Damage               float64 `json:"Damage"bson:"damage"`
+		Headshots            float64 `json:"Headshots"bson:"headshots"`
 		ID                   int     `json:"Id"bson:"id"`
-		Kd                   int     `json:"Kd"bson:"kd"`
-		KillRank             int     `json:"KillRank"bson:"kill_rank"`
-		KillRating           int     `json:"KillRating"bson:"kill_rating"`
-		KillRatingChange     int     `json:"KillRatingChange"bson:"kill_rating_change"`
-		KillRatingRankChange int     `json:"KillRatingRankChange"bson:"kill_rating_rank_rhange"`
-		Kills                int     `json:"Kills"bson:"kills"`
-		Match                int     `json:"Match"bson:"match"`
+		Kd                   float64 `json:"Kd"bson:"kd"`
+		KillRank             float64 `json:"KillRank"bson:"kill_rank"`
+		KillRating           float64 `json:"KillRating"bson:"kill_rating"`
+		KillRatingChange     float64 `json:"KillRatingChange"bson:"kill_rating_change"`
+		KillRatingRankChange float64 `json:"KillRatingRankChange"bson:"kill_rating_rank_rhange"`
+		Kills                float64 `json:"Kills"bson:"kills"`
+		Match                float64 `json:"Match"bson:"match"`
 		MatchDisplay         string  `json:"MatchDisplay"bson:"match_display"`
 		MoveDistance         float64 `json:"MoveDistance"bson:"move_distance"`
-		Rating               int     `json:"Rating"bson:"rating"`
+		Rating               float64 `json:"Rating"bson:"rating"`
 		RatingChange         float64 `json:"RatingChange"bson:"rating_change"`
-		RatingRank           int     `json:"RatingRank"bson:"rating_rank"`
-		RatingRankChange     int     `json:"RatingRankChange"bson:"rating_rank_change"`
-		Region               int     `json:"Region"bson:"region"`
+		RatingRank           float64 `json:"RatingRank"bson:"rating_rank"`
+		RatingRankChange     float64 `json:"RatingRankChange"bson:"rating_rank_change"`
+		Region               float64 `json:"Region"bson:"region"`
 		RegionDisplay        string  `json:"RegionDisplay"bson:"region_display"`
-		Rounds               int     `json:"Rounds"bson:"rounds"`
-		Season               int     `json:"Season"bson:"season"`
+		Rounds               float64 `json:"Rounds"bson:"rounds"`
+		Season               float64 `json:"Season"bson:"season"`
 		SeasonDisplay        string  `json:"SeasonDisplay"bson:"season_display"`
 		TimeSurvived         float64 `json:"TimeSurvived"bson:"time_survived"`
 		Top10                int     `json:"Top10"bson:"top10"`
 		Updated              string  `json:"Updated"bson:"updated"`
 		UpdatedJS            string  `json:"UpdatedJS"bson:"updated_js"`
-		WinRank              int     `json:"WinRank"bson:"win_rank"`
-		WinRating            int     `json:"WinRating"bson:"win_rating"`
-		WinRatingChange      int     `json:"WinRatingChange"bson:"win_rating_change"`
-		WinRatingRankChange  int     `json:"WinRatingRankChange"bson:"win_rating_rank_change"`
+		WinRank              float64 `json:"WinRank"bson:"win_rank"`
+		WinRating            float64 `json:"WinRating"bson:"win_rating"`
+		WinRatingChange      float64 `json:"WinRatingChange"bson:"win_rating_change"`
+		WinRatingRankChange  float64 `json:"WinRatingRankChange"bson:"win_rating_rank_change"`
 		Wins                 int     `json:"Wins"bson:"wins"`
 	} `json:"MatchHistory"bson:"match_history"`
 	Platform      int    `json:"Platform"bson:"platform"`
